pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "atlasvras"
version = "1.1.1"

authors = [
    {name = "Heloise Stevance", email = "hfstevance@gmail.com"},
]
description = "ATLAS Virtual Research Assistants"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["astronomy", "ATLAS VRAs"]
license = {text = "BSD-3-Clause"}
dependencies = [
    "numpy",
    "pandas",
    "matplotlib",
    "astropy",
    "requests",
    "pyyaml",
    "tqdm",
    "joblib",
    "scikit-learn",
    "dustmaps",
    "atlasapiclient",

    # Slack bot
    "slack_sdk",
    "slack_bolt",

    # Testing
    "pytest",
    "pytest-cov",
    'importlib-metadata; python_version>="3.8"'
]

# See some of the stuff I have in the hoki .toml file 
[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow",
    "integration: mark tests as requiring additional infrastructure",
    "serial",
]

[project.optional-dependencies]
test = ["pytest"]  
docs = ["sphinx==7.4.7", "sphinx_rtd_theme==3.0.2"]

#[project.optional-dependencies]
#pdf = ["ReportLab>=1.2", "RXP"]
#rest = ["docutils>=0.3", "pack ==1.1, ==1.3"]

#[project.scripts]
#my-script = "my_package.module:function"

# ... other project metadata fields as specified in:
#     https://packaging.python.org/en/latest/specifications/declaring-project-metadata/

